requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-braces==1.14.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.9
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
oauthlib==3.1.1
packaging==20.3
pep517==0.8.2
pep8==1.7.1
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
